% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{bernoulli}
\alias{bernoulli}
\title{Bernoulli Distribution}
\usage{
bernoulli(prob, dim = NULL)
}
\arguments{
\item{prob}{Probability of success (numeric or \code{gretaR_array}, 0 to 1).}

\item{dim}{Integer vector of dimensions (default scalar).}
}
\value{
A \code{gretaR_array}.
}
\description{
Create a Bernoulli-distributed variable with the specified
probability of success. Support is on \code{{0, 1}}.
}
\examples{
\dontrun{
z <- bernoulli(0.5)
}
}
