% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{binomial}
\alias{binomial}
\title{Binomial Distribution}
\usage{
binomial(size, prob, dim = NULL)
}
\arguments{
\item{size}{Number of trials (positive integer or \code{gretaR_array}).}

\item{prob}{Probability of success per trial (numeric or \code{gretaR_array}, 0 to 1).}

\item{dim}{Integer vector of dimensions (default scalar).}
}
\value{
A \code{gretaR_array}.
}
\description{
Create a binomially-distributed variable with the specified
number of trials and probability of success per trial.
}
\examples{
\dontrun{
y <- binomial(size = 10, prob = 0.3)
}
}
