% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\docType{data}
\name{.gretaR_reason_codes}
\alias{.gretaR_reason_codes}
\title{Reason codes for a gretaR refusal}
\format{
An object of class \code{character} of length 5.
}
\usage{
.gretaR_reason_codes
}
\description{
Reason codes for a gretaR refusal
}
\keyword{internal}
