% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{exponential}
\alias{exponential}
\title{Exponential Distribution}
\usage{
exponential(rate = 1, dim = NULL, truncation = NULL)
}
\arguments{
\item{rate}{Rate parameter (positive numeric or \code{gretaR_array}).}

\item{dim}{Integer vector of dimensions (default scalar).}

\item{truncation}{Optional length-2 numeric vector \code{c(lower, upper)}
specifying truncation bounds. Default \code{NULL} (no truncation).}
}
\value{
A \code{gretaR_array}.
}
\description{
Create an exponentially-distributed variable with the specified
rate. Support is on the positive reals.
}
\examples{
\dontrun{
lambda <- exponential(rate = 1)
}
}
