% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gamma_dist}
\alias{gamma_dist}
\title{Gamma Distribution}
\usage{
gamma_dist(shape, rate, dim = NULL)
}
\arguments{
\item{shape}{Shape parameter (positive numeric or \code{gretaR_array}).}

\item{rate}{Rate parameter (positive numeric or \code{gretaR_array}).}

\item{dim}{Integer vector of dimensions (default scalar).}
}
\value{
A \code{gretaR_array}.
}
\description{
Create a gamma-distributed variable with the specified shape
and rate. Support is on the positive reals. Named \code{gamma_dist} to
avoid conflict with \code{base::gamma}.
}
\examples{
\dontrun{
tau <- gamma_dist(shape = 2, rate = 1)
}
}
