% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{half_normal}
\alias{half_normal}
\title{Half-Normal Distribution}
\usage{
half_normal(sd = 1, dim = NULL)
}
\arguments{
\item{sd}{Scale parameter (positive numeric or \code{gretaR_array}).}

\item{dim}{Integer vector of dimensions (default scalar).}
}
\value{
A \code{gretaR_array} with support on the positive reals.
}
\description{
Create a half-normal-distributed variable. The half-normal is
the absolute value of a normal distribution, with support on the positive
reals.
}
\examples{
\dontrun{
sigma <- half_normal(1)
}
}
