% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{joint_density}
\alias{joint_density}
\title{Get the Log Joint Density Function}
\usage{
joint_density(model)
}
\arguments{
\item{model}{A \code{gretaR_model} object created by \code{\link{model}}.}
}
\value{
A function \code{f(theta)} that takes a 1-D torch tensor of
unconstrained parameter values and returns a scalar torch tensor (the
log joint density).
}
\description{
Extract a torch-compatible function that evaluates the log
joint density of a compiled \code{gretaR_model} at a given parameter
vector in unconstrained space.
}
\examples{
\dontrun{
mu <- normal(0, 10)
y <- as_data(rnorm(50, 3))
distribution(y) <- normal(mu, 1)
m <- model(mu)
ld <- joint_density(m)
ld(torch::torch_zeros(1))
}
}
