% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{lognormal}
\alias{lognormal}
\title{Log-Normal Distribution}
\usage{
lognormal(meanlog = 0, sdlog = 1, dim = NULL)
}
\arguments{
\item{meanlog}{Mean of the log-scale distribution.}

\item{sdlog}{Standard deviation on the log scale (positive).}

\item{dim}{Dimensions.}
}
\value{
A \code{gretaR_array} with support on the positive reals.
}
\description{
A variable whose logarithm is normally distributed.
}
\examples{
\dontrun{
x <- lognormal(0, 1)
}
}
