% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{mean.gretaR_array}
\alias{mean.gretaR_array}
\title{Mean of a gretaR_array}
\usage{
\method{mean}{gretaR_array}(x, ...)
}
\arguments{
\item{x}{A \code{gretaR_array}.}

\item{...}{Ignored.}
}
\value{
A scalar \code{gretaR_array}.
}
\description{
Build an operation node that reduces a \code{gretaR_array} to a scalar by
averaging over all elements.
}
