% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{normal}
\alias{normal}
\title{Normal Distribution}
\usage{
normal(mean = 0, sd = 1, dim = NULL)
}
\arguments{
\item{mean}{Mean of the distribution (numeric or \code{gretaR_array}).}

\item{sd}{Standard deviation (numeric or \code{gretaR_array}, positive).}

\item{dim}{Integer vector of dimensions (default scalar).}
}
\value{
A \code{gretaR_array} representing a normally-distributed variable.
}
\description{
Create a normally-distributed variable with the specified mean
and standard deviation. Support is on the entire real line.
}
\examples{
\dontrun{
x <- normal(0, 1)
beta <- normal(0, 5, dim = c(3, 1))
}
}
