% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{sum.gretaR_array}
\alias{sum.gretaR_array}
\title{Sum of a gretaR_array}
\usage{
\method{sum}{gretaR_array}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A \code{gretaR_array}.}

\item{...}{Ignored.}

\item{na.rm}{Ignored (no NAs in torch).}
}
\value{
A scalar \code{gretaR_array}.
}
\description{
Build an operation node that reduces a \code{gretaR_array} to a scalar by
summing over all elements.
}
