% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{uniform}
\alias{uniform}
\title{Uniform Distribution}
\usage{
uniform(lower = 0, upper = 1, dim = NULL)
}
\arguments{
\item{lower}{Lower bound (numeric).}

\item{upper}{Upper bound (numeric).}

\item{dim}{Integer vector of dimensions (default scalar).}
}
\value{
A \code{gretaR_array}.
}
\description{
Create a uniformly-distributed variable on the interval
\code{[lower, upper]}.
}
\examples{
\dontrun{
p <- uniform(0, 1)
}
}
