% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gretaR-package.R
\docType{package}
\name{gretaR-package}
\alias{gretaR}
\alias{gretaR-package}
\title{gretaR: Bayesian Statistical Modelling with Torch}
\description{
A probabilistic programming package for Bayesian statistical modelling in R
using native R syntax. Models are defined interactively with \code{gretaR_array}
objects, then compiled to torch tensors for GPU-accelerated HMC and NUTS
inference.
}
\section{Core workflow}{

\enumerate{
\item Wrap observed data with \code{\link[=as_data]{as_data()}}
\item Define priors using distribution functions (e.g., \code{\link[=normal]{normal()}}, \code{\link[=gamma_dist]{gamma_dist()}})
\item Define the model structure using standard R operations
\item Assign a likelihood with \code{\link[=distribution]{distribution()}}
\item Create a model with \code{\link[=model]{model()}}
\item Draw samples with \code{\link[=mcmc]{mcmc()}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/max578/gretaR}
  \item Report bugs at \url{https://github.com/max578/gretaR/issues}
}

}
\author{
\strong{Maintainer}: Max Moldovan \email{max.moldovan@adelaide.edu.au} (\href{https://orcid.org/0000-0001-9680-8474}{ORCID})

}
\keyword{internal}
